Fix incoherent tree after traversing.
When we traversed the tree, it was no longer fully consistent. For
example, with the following tree:

    t = root $ branch 'A' [branch 'B' [branch 'C' [leaf 'D'],
         branch 'E' [leaf 'F',leaf 'G']],leaf 'H']

... we would have obtained Nothing if we tried:

    (firstChild >=> firstChild >=> right >=> firstChild >=> up >=> up) $ t

... even though we should have gotten the subtree with 'B' at the root.
This would start to be a big issue once we want to start manipulating
the tree.

It is now fixed by keeping track of context during traversal. When we go
to the left sibling, for example, we also make sure that the focus of
the parent node travels along with us. This ensures that all references
stay fully updated.

Referential transparency makes things a bit difficult here, so I figure
that we should make things a bit easier still.