Keep track of parent nodes correctly.
Before, keeping every reference of the context of a node was a bit messy
and opaque. This update solves bugs that were introduced as a result.

A node's context is now directly integrated in a `Node`. Although this
does make us lose the guarantee on the type-level that a root node has
neither a parent nor any siblings, it makes the code clearer. It
probably won't lead to problems, since coherence of the root node is
almost a natural consequence anyway; see the function `left`, for
example: if the node has no parent, going to the left sibling would also
be blocked.

Moreover, some variable names have been made more succinct. I like to do
as little namespacing within a module as possible: that should be
handled by the module system.