Fix another traversal mistake.
The node you leave behind when travelling laterally used to contain an
old reference. This bug surfaced because of the following snippet:

    let t = branch () [leaf 2, leaf 3]
    withTree t $ addFront (leaf 1) >=> lastChild

Clearly, this should equal `branch () [leaf 1, leaf 2, leaf 3]`, but
instead it returned `branch () [leaf 2, leaf 3]`, because information
got lost once you travelled to the `lastChild`. This is fixed now.

Since this is so tricky, I should probably find a way to make it
entirely obvious. At the very least, unit tests should be included.